Create the following modules Create_person create_student create_teacher
